backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
aiohttp
jose
cloudflare
boto3 
pandas
google-generativeai
openpyxl
xlsxwriter